[refactor]: 비활성화 페이지 컴포넌트 분리
